Fix double slash for auth transfer
The definition of the variable bind_conf_dir: "/etc/named/" for Redhat
and bind_conf_dir: "/etc/bind/" for Debian with a trailing slash added
to the definition of bind_auth_file: "{{ bind_conf_dir }}/{{ auth_file
}}" for both families causes a double slash while templating named.conf.

It does not generate an error but can be easily avoided by removing the
trailing slash from bind_conf_dir. This is what this modification does.